Move all Cogs traits into one file
