+++ lib/Traits/TraitsClasses/CogCatalogItemTraits.ts
@@ -1,14 +1,30 @@
+import objectTrait from "../Decorators/objectTrait";
+import primitiveTrait from "../Decorators/primitiveTrait";
+import ModelTraits from "../ModelTraits";
+import { traitClass } from "../Trait";
 import mixTraits from "../mixTraits";
 import CatalogMemberTraits from "./CatalogMemberTraits";
+import ImageryProviderTraits from "./ImageryProviderTraits";
 import LayerOrderingTraits from "./LayerOrderingTraits";
 import LegendOwnerTraits from "./LegendOwnerTraits";
 import MappableTraits from "./MappableTraits";
-import ImageryProviderTraits from "./ImageryProviderTraits";
 import UrlTraits from "./UrlTraits";
-import { traitClass } from "../Trait";
-import primitiveTrait from "../Decorators/primitiveTrait";
-import objectTrait from "../Decorators/objectTrait";
-import { CogRenderOptionsTraits } from "./CogRenderOptionsTraits";
+
+export class CogRenderOptionsTraits extends ModelTraits {
+  @primitiveTrait({
+    type: "number",
+    name: "No Data Value",
+    description: "No data value, default read from tiff meta"
+  })
+  nodata?: number;
+
+  @primitiveTrait({
+    type: "boolean",
+    name: "Convert to RGB",
+    description: "Try to render multi band cog to RGB, priority 1"
+  })
+  convertToRGB?: boolean;
+}
 
 @traitClass({
   description:
